+++ Seguim.Netcore.Store.Domain/StoreContext/Entities/OrderItem.cs
@@ -1,4 +1,4 @@
-namespace Seguim.Netcore.Store.Domain.StoreContext
+namespace Seguim.Netcore.Store.Domain.StoreContext.Entities
 {
     public class OrderItem
     {
