+++ Seguim.Netcore.Store.Domain/StoreContext/Entities/Product.cs
@@ -1,4 +1,4 @@
-namespace Seguim.Netcore.Store.Domain.StoreContext
+namespace Seguim.Netcore.Store.Domain.StoreContext.Entities
 {
     public class Product
     {
